Templates are now returned by all handlers.
A new type, Page, has been added, which will
be returned by all handlers.

A page is an interface with an Execution
method. This method will be called along with
the writter to write the page to the client.

So far, a page can either be a Template or
a Redirect.

This changes aims to make testing easier since
only the "context" (the template args, or the
redirection URL) needs to be tested.